Add partial `Map` support to assembler
Allow Assembler customizations to replace Object instances with Map
without affecting the rest of its functionality.  It makes it much more
efficient when deserializing large "objects" that are functionally
lookup tables, with minimal customization.

The mechanism for determining when replacement with Map is used is
out-of-scope.  There are no new APIs added.